move fns from experiment helpers to core
